backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.11.1
pydantic-settings==2.0.3
python-dotenv==1.0.0
openai==0.28.1
pytest==7.4.0
httpx==0.25.0
pyyaml==6.0